+++ magda-migrator-content-db/sql/V1_4__update_config.sql
@@ -1,7 +1,0 @@
-UPDATE public.content
-SET "id" = "header/logo"
-WHERE "id" = "logo";
-
-UPDATE public.content
-SET "id" = "header/logo-mobile"
-WHERE "id" = "logo-mobile";