+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## 0.0.50
 
+-   Make header menu items come from content api
 -   Changed some existing content ids to group things better
 -   Reorganised content api to require less number of requests to get small JSON content items
 -   Email Templates are configurable from content API